+++ apps/settings/lib/Settings/Personal/Security.php
@@ -27,25 +27,8 @@
 
 namespace OCA\Settings\Personal;
 
-
-use function array_filter;
-use function array_map;
-use function is_null;
-use OC\Authentication\Exceptions\InvalidTokenException;
-use OC\Authentication\Token\INamedToken;
-use OC\Authentication\Token\IProvider as IAuthTokenProvider;
-use OC\Authentication\Token\IToken;
-use OC\Authentication\TwoFactorAuth\Manager as TwoFactorManager;
-use OC\Authentication\TwoFactorAuth\ProviderLoader;
 use OCP\AppFramework\Http\TemplateResponse;
-use OCP\Authentication\TwoFactorAuth\IProvider;
-use OCP\Authentication\TwoFactorAuth\IProvidesPersonalSettings;
-use OCP\IConfig;
-use OCP\IInitialStateService;
-use OCP\ISession;
 use OCP\IUserManager;
-use OCP\IUserSession;
-use OCP\Session\Exceptions\SessionNotAvailableException;
 use OCP\Settings\ISettings;
 
 class Security implements ISettings {
@@ -53,28 +36,13 @@ class Security implements ISettings {
 	/** @var IUserManager */
 	private $userManager;
 
-	/** @var ProviderLoader */
-	private $providerLoader;
-
-	/** @var IUserSession */
-	private $userSession;
-
 	/** @var string|null */
 	private $uid;
 
-	/** @var IConfig */
-	private $config;
-
 	public function __construct(IUserManager $userManager,
-								ProviderLoader $providerLoader,
-								IUserSession $userSession,
-								IConfig $config,
 								?string $UserId) {
 		$this->userManager = $userManager;
-		$this->providerLoader = $providerLoader;
-		$this->userSession = $userSession;
 		$this->uid = $UserId;
-		$this->config = $config;
 	}
 
 	public function getForm(): TemplateResponse {
@@ -86,10 +54,7 @@ class Security implements ISettings {
 
 		return new TemplateResponse('settings', 'settings/personal/security', [
 			'passwordChangeSupported' => $passwordChangeSupported,
-			'twoFactorProviderData' => $this->getTwoFactorProviderData(),
-			'themedark' => $this->config->getUserValue($this->uid, 'accessibility', 'theme', false)
 		]);
-
 	}
 
 	public function getSection(): string {
@@ -98,24 +63,5 @@ class Security implements ISettings {
 
 	public function getPriority(): int {
 		return 10;
-	}
-
-	private function getTwoFactorProviderData(): array {
-		$user = $this->userSession->getUser();
-		if (is_null($user)) {
-			// Actually impossible, but still …
-			return [];
-		}
-
-		return [
-			'providers' => array_map(function (IProvidesPersonalSettings $provider) use ($user) {
-				return [
-					'provider' => $provider,
-					'settings' => $provider->getPersonalSettings($user)
-				];
-			}, array_filter($this->providerLoader->getProviders($user), function (IProvider $provider) {
-				return $provider instanceof IProvidesPersonalSettings;
-			}))
-		];
 	}
 }
