+++ apps/settings/templates/settings/personal/security.php
@@ -59,48 +59,3 @@ if($_['passwordChangeSupported']) {
 	<span class="msg"></span>
 </div>
 <?php } ?>
-
-<div id="two-factor-auth" class="section">
-	<h2><?php p($l->t('Two-Factor Authentication'));?></h2>
-	<a target="_blank" rel="noreferrer noopener" class="icon-info"
-	   title="<?php p($l->t('Open documentation'));?>"
-	   href="<?php p(link_to_docs('user-2fa')); ?>"></a>
-	<p class="settings-hint"><?php p($l->t('Use a second factor besides your password to increase security for your account.'));?></p>
-	<ul>
-	<?php foreach ($_['twoFactorProviderData']['providers'] as $data) { ?>
-		<li>
-			<?php
-
-			/** @var \OCP\Authentication\TwoFactorAuth\IProvidesPersonalSettings $provider */
-			$provider = $data['provider'];
-			//Handle 2FA provider icons and theme
-			if ($provider instanceof \OCP\Authentication\TwoFactorAuth\IProvidesIcons) {
-				if ($_['themedark']) {
-					$icon = $provider->getLightIcon();
-				}
-				else {
-					$icon = $provider->getDarkIcon();
-				}
-				//fallback icon if the 2factor provider doesn't provide an icon.
-			} else {
-				if ($_['themedark']) {
-					$icon = image_path('core', 'actions/password-white.svg');
-				}
-				else {
-					$icon = image_path('core', 'actions/password.svg');
-				}
-
-			}
-			/** @var \OCP\Authentication\TwoFactorAuth\IPersonalProviderSettings $settings */
-			$settings = $data['settings'];
-			?>
-			<h3>
-				<img class="two-factor-provider-settings-icon" src="<?php p($icon) ?>" alt="">
-				<?php p($provider->getDisplayName()) ?>
-			</h3>
-			<?php print_unescaped($settings->getBody()->fetchPage()) ?>
-		</li>
-	<?php } ?>
-	</ul>
-</div>
-
